requirements/prod.txt
#aiogram==3.0.0b7
asyncpg>=0.27.0
pandas>=2.2.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
#alembic>=1.7.0
#redis>=4.3.0
#pydantic-settings>=2.0.0
sqlalchemy==2.0.39
psycopg2-binary>=2.9.10
uvicorn>=0.34.0
fastapi>=0.115.11
pydantic_settings>=2.8.1
werkzeug>=3.1.3
